requirements.txt
altair==4.0.1
appnope==0.1.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
base58==2.0.0
beautifulsoup4==4.8.2
bleach==3.1.1
blinker==1.4
boto3==1.12.6
botocore==1.15.6
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
defusedxml==0.6.0
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
fake-useragent==0.1.11
future==0.18.2
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-vision==0.42.0
Google-Search-API==1.1.14
googleapis-common-protos==1.51.0
grpcio==1.27.2
idna==2.9
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
jupyter-client==6.0.0
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
multidict==4.7.5
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.2.1
Pygments==2.5.2
pyrsistent==0.15.7
PySocks==1.7.1
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.3
selenium==2.53.6
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0
streamlit==0.56.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.0
toolz==0.10.0
tornado==5.1.1
traitlets==4.3.3
tweepy==3.8.0
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.8
validators==0.14.2
vcrpy==4.0.2
watchdog==0.10.2
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.0
yarl==1.4.2
